backend/pyproject.toml
[project]
name = "shitposter-backend"
version = "0.1.0"
description = "FastAPI backend for Shitposter"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.17.0",
  "arq>=0.26.0",
  "asyncpg>=0.31.0",
  "authlib>=1.6.0",
  "bcrypt>=4.1.0,<5.0.0",
  "cryptography>=46.0.0",
  "email-validator>=2.3.0",
  "fastapi>=0.121.0",
  "httpx>=0.28.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.12.0",
  "python-multipart>=0.0.20",
  "redis>=5.0.0,<6.0.0",
  "sentry-sdk[fastapi]>=2.47.0",
  "sqlalchemy[asyncio]>=2.0.44",
  "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
  "asgi-lifespan>=2.1.0",
  "aiosqlite>=0.21.0",
  "mypy>=1.18.0",
  "pytest>=9.0.0",
  "pytest-asyncio>=1.2.0",
  "ruff>=0.14.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.mypy]
python_version = "3.12"
plugins = ["pydantic.mypy"]
strict = true
warn_unused_configs = true
exclude = ["alembic"]
